✨ feat(server): validate catalog on PUT/PATCH/seed/runtime apply
Shared agent_catalog helper runs validate_catalog across config
write path, builtin seed, and runtime candidate validation.
